Remove padding from DropdownListItem component
The padding of the DropdownListItem was causing issues in cases where we were using a ButtonDropdownMenu in Course Planner. Each list item of the ButtonDropdownMenu contains a ButtonDropdownMenuItem, which each have padding. With the DropdownListItem component also having padding, if the user clicked on the DropdownListItem's padding, the menu would collapse without triggering a click event on the ButtonDropdownMenuItem, causing confusion for the user.