Replace modal onClick handler with a mouseDown handler
A flag was added to indicate whether there was a mouseDown event on the modal. If there was, it sets the flag to true so that if the user then drags the mouse outside the modal and releases the mouse, the modal does not close (since there is a conditional checking that the flag is false before calling the closeHandler). The flag is then set back to false so that the logic can continue for subsequent clicks. The reason why a state value wasn't used was to ensure that re-rendering of the modal didn't happen every time the state value was updated.