Add Footer component
This component is modeled after the existing Header component in terms of format. It accepts optional justify-content and background color props, and it will display the contents that are passed to it via the children prop. The background color is transparent by default, and justify-content is set to space-between by default, but these can be changed depending on the needs of the user.