Update Button component's onClick property type
The type for onClick was updated to accommodate providing async functions as a value for onClick. This is an attempt to fix the linter error in Course Planner: "Promise-returning function provided to attribute where a void return was expected" for the onClick value of the submit button in the modals.