Add padding-bottom to PageBody
The purpose of adding a padding to PageBody is to solve for an issue in Course Planner in which the footer covers the bottom of a table making the information in the table inaccessible to the user. By adding some padding to the bottom of PageBody, we are making it so that the footer does not obfuscate any information.