Add position prop
There are some cases where we don't need the footer to float, so we
should be allowed to pass in a different value when needed. The default
of "sticky" will emulate the previous "fixed" behavior, but will
unsticky the footer when the user reaches the end of the page content.
That should resolve the issue with the footer covering up useful page
content.