Override DefaultTheme in package
Previously, our theme override was not being included in the distributed
package, so we were ending up with the standard styled-components
`DefaultTheme` type in our consuming packages. This change will add the
theme declaration to our compiled output, and should mean that we won't
need to duplicate the theme file in our projects anymore.