Convert theme file to module
It seems that the DefaultTheme override in the d.ts file was causing
problems in the build phase -- mainly that the declarations were not
being carried over into the final output. To resolve, we're removing the
"d." from the filename, and exporting an empty module from the new file.
Importing that empty module into our index barrel file seems to be
enough to pick up the declarations and package them into the final
output.