Update various packages to solve typecheck issues
There was originally an issue with using MemoryRouter from the react-router-dom package. After updating the react-router dom package, we found that the types are now integrated with this package, so @types/react-router-dom was removed. After that, we were continually getting errors related to the node_modules files. We found that we had to update typescript along with jsdom. The react-styleguidist package was also updated due to warnings of incompatibility with React 18.