Fix linter errors associated with GlobalMessage component
The id property is not necessary in this context, because we are not trying to associate a label with a form. In instances that we use GlobalMessage, we are not using the id. The variant was marked as an optional property, since we provide a default value for it. A default value of null for the forwardRef is provided.