Update getRecursiveProps function's type definitions
This is in response to getting type error: "Argument of type 'string' is not assignable to parameter of type 'never'" when trying to access the second Mark One Theme property using "fromTheme". When using "Extract" in this way, it could not properly recognize which keys were valid. This way, we are checking that the property is included in the valid theme values and is also a key of the object the current key is within.