Fix Form component
When trying to use the Form component in the makerspace project, we get the following error message: "Type '{ children: Element[]; id: string; label: string; submitHandler: () => void; }' is not assignable to type 'IntrinsicAttributes & FormProps'.
  Property 'children' does not exist on type 'IntrinsicAttributes & FormProps'."
This issue was fixed in the past for mark one components by explicitly passing down a children prop.