Declare ModalBody and ModalFooter types
PropsWithChildren was included within the definition of the types for the ModalBody and ModalFooter. These changes are in response to getting a type error message when trying to use these components in the Makerspace project related to the "children" prop. (Type '{ children: Element; }' has no properties in common with type 'IntrinsicAttributes & ExecutionProps & RefAttributes<IStyledComponent<"web", Substitute<DetailedHTMLProps<HTMLAttributes<HTMLDivElement>, HTMLDivElement>, { ...; }>>>'.ts(2559)