Include typing of PropsWithChildren in the definition of ModalFooter and ModalBody
As a result of our dev standup discussion today, we found that we must include the PropsWithChildren type within the definition of the Modal components in addition to providing this typing in the ModalFooter and ModalBody types. Otherwise, we get a type issue when using these components that 'children' is missing as a prop for these components.