Don't check test coverage
This is still running the coverage report and giving us the output after
the tests run, but we don't need to fail the test run if we don't meet
the 80% threshold. Philosophically, test coverage is not the best
indicator of code quality or test quality, and we'd rather leave it to
our best judgement about what needs to be tested, rather than writing
tests just to make the number green.